refactor(product page): Split product model server code from client code
