+++ frontend/templates/product/MetaTags.tsx
@@ -1,6 +1,6 @@
 import { useAppContext } from '@ifixit/app';
 import { parseItemcode } from '@ifixit/helpers';
-import { Product, ProductVariant } from '@models/product';
+import type { Product, ProductVariant } from '@models/product.server';
 import Head from 'next/head';
 import React from 'react';
 import { jsonLdScriptProps } from 'react-schemaorg';
