+++ frontend/templates/product/hooks/useInternationalBuyBox.tsx
@@ -1,10 +1,9 @@
-import { Product } from '@models/product';
-import { useQuery } from '@tanstack/react-query';
-import { useState } from 'react';
 import { useIFixitApiClient } from '@ifixit/ifixit-api-client';
-import { useSelectedVariant } from './useSelectedVariant';
-import { getBuyBoxForProduct } from '@lib/ifixit-api/international-buy-box';
 import { useExpiringLocalPreference } from '@ifixit/ui';
+import { getBuyBoxForProduct } from '@lib/ifixit-api/international-buy-box';
+import type { Product } from '@models/product.server';
+import { useQuery } from '@tanstack/react-query';
+import { useSelectedVariant } from './useSelectedVariant';
 
 const buyBoxKey = (productcode: string) => ['buy-box-api', productcode];
