+++ frontend/templates/product/hooks/useIsProductForSale.ts
@@ -1,5 +1,5 @@
 import { useAuthenticatedUser } from '@ifixit/auth-sdk';
-import { Product } from '@models/product';
+import type { Product } from '@models/product.server';
 
 export function useIsProductForSale(product: Product): boolean {
    const user = useAuthenticatedUser();
