+++ frontend/templates/product/hooks/useProductTemplateProps.ts
@@ -1,7 +1,7 @@
 import { WithProvidersProps } from '@components/common';
 import type { WithLayoutProps } from '@layouts/default/server';
 import { useServerSideProps } from '@lib/server-side-props';
-import { Product } from '@models/product';
+import type { Product } from '@models/product.server';
 
 export type ProductTemplateProps = WithProvidersProps<
    WithLayoutProps<{
