+++ frontend/templates/product/sections/ProductSection/ProductDescription.tsx
@@ -5,7 +5,7 @@ import {
    faTriangleExclamation,
 } from '@fortawesome/pro-solid-svg-icons';
 import { FaIcon } from '@ifixit/icons';
-import { Product, ProductVariant } from '@models/product';
+import type { Product, ProductVariant } from '@models/product.server';
 
 export type ProductDescriptionProps = {
    product: Product;
