+++ frontend/templates/product/sections/ServiceValuePropositionSection/index.tsx
@@ -1,4 +1,4 @@
-import { ProductVariant } from '@models/product';
+import type { ProductVariant } from '@models/product.server';
 import { Heading, Stack, Text, VStack } from '@chakra-ui/react';
 import React from 'react';
 import { List, ListIcon, ListItem } from '@chakra-ui/react';
