+++ frontend/templates/product/server.tsx
@@ -7,7 +7,7 @@ import { ifixitOriginFromHost } from '@helpers/path-helpers';
 import { invariant } from '@ifixit/helpers';
 import { urlFromContext } from '@ifixit/helpers/nextjs';
 import { getLayoutServerSideProps } from '@layouts/default/server';
-import { findProduct } from '@models/product';
+import { findProduct } from '@models/product.server';
 import { GetServerSideProps } from 'next';
 import { ProductTemplateProps } from './hooks/useProductTemplateProps';
